Created 2sorted_linkedlist.cpp
Approach: Here we have First created the basic structure of a linkedlist. nd then we push the value of 2 linkedlist in an array using push_back function. after that we use a basic sort() function to sort it. This is not enough, now we have to take an extra space. lets say a dummy node where we'll store our list. Now lets print the sorted list. 
That's it.