Create Palindrome_linkedlist.cpp
So here, first we create the linkedlist basic structure and later we create a function to chck if it is palindrome or not. For this, we will intialize a stack and we will push all the elements of list into it. intialize a variable lets say c, which will store the value of the stack's top and now we chck if it is equal to head's data or not. if equal it continues with the rest of linkedlist if not true it returns false.