Create 3. Merge_k_sorted.cpp
So the approach is, first a list structure is created. we create get min function to find the minimum value among the heads of linkedlist. It iterates through the vector to find the node with min. value. if found, it updates index to point next node in tht list. returns node with min val.
Now, we'll create merge k sorted list. We create a dummy node and a tail to keep track of last node in merged list. we repeatedly call getmin function to find the min val. Each min node is appended to merged list and tail ptr is updated. Once it becomes null min_ var becomes null ptr. finally we merge list starting frm nodes after dummy node.