Created rearrange_even_odd
so here is the approach. first create the basic structure of a linkedlist. then we'll create a function named rearrangeEvenOdd. so the base case for this will be if the node is nullptr or has only single node then, it returns as no rearrangment required.
for the rest, we'll create 2 dummy node for odd nd even head for odd nd even positioned head. pointers are initialized fr the same.
now we'll traverse the list if the current position is odd node is added to odd. same is done for even. nd we will keep moving the curr pointer to the next node nd the next position is incremented.
after traversing whole list, we will make odd next to evenhead next nd the end of the list will be nullptr.