Create 5.rearrange_list.cpp
Here is the approach. first create a structure of linkedlist and then a reverse linkedlist function also a function to print the list. now lets create a function to rearrange the nodes. so for this we will create 2 pointers: slow (stores head val) and fast(slow->next val)
now we'll create 2 nodes head1 nd head2 having value head nd slow->next respectively to split the linkedlist into 2 halves. we will reverse the second half i.e. head2.
create a dummy node store the value in current node. we will traverse head1 nd head2 then when we find head1 we push the value in the list nd same for head2